Rework the explanation of why we need to qualify “blockchain”
I don’t think the word has become debased. It is thrown around a lot by
marketers, certainly, but unlike, say, “agile”, which has come to mean a
huge variety of things, and is hence arguably meaningless, “blockchain”
still means “mutual distributed ledger”. The abuse of the word consists more
in adding “blockchain” to a phrase to make it sound appealing, in the way
one might add “quantum”, without diluting the meaning of the word.

However, the word has become much vaguer thanks to the enormous variety of
blockchain designs introduced in recent times, so the idea that it needs
qualifying is bang on.

I thought this point worth explaining because the rewrite is not owing to my
being in sympathy with marketers! Rather it’s an interesting coincidence
that I agree with the point you’re making, but not the reasons.