Update 07smart-contracts-solidity.asciidoc
You were incorrectly stating that if a transaction is successful (processed without errors) that all the global state changes it performed are accepted. This is not true when a transaction is sent from an EOA to a contract A, and that contract A calls another contract B using a Solidity method such as .call(...), which will return false on error. And this error could occur after some variables of B were changed. See https://github.com/Luker501/SmartContractInteractions for more information.